Tests (#7)
* Add files via upload

Updated diag_peels, inlined all functions in utilities and kernels to fix the linking issues in the tests, and added a tests folder. Plan is to flesh out tests for kernel and utiltites before moving onto tests for the solvers.

* Add files via upload

Updated diag_peels, inlined all functions in utilities and kernels to fix the linking issues in the tests, and added a tests folder. Plan is to flesh out tests for kernel and utiltites before moving onto tests for the solvers.

* Update sparse_matrix.hpp

* Update CMakeLists.txt

* Add files via upload

* Update test_kernels.cpp

* Update test_utilities.cpp

* Update cg.hpp

* Update gmres.hpp

* Update bicgstab.hpp

* Update CMakeLists.txt

* Update test_solvers.cpp

Added another test for BiCGSTAB

* small tweaks and clang-formatting; add precond. residual fixes for BiCGSTAB and update figs

---------

Co-authored-by: BallasticMonkey <62638445+BallasticMonkey@users.noreply.github.com>